Make basic client more flexible
We found a case that we have an existing launch and only need to create rootItem under that launch. Thus make a setter for launchID property so we can set it.

Also the finishRootItem somehow set the description to empty string. Make it to accept a string and set it.